NotificationCard: Make most notifications stand out a lot less
Instead of placing a card within a card, thus taking not only a lot of space,
but putting emphasis on that is less important, blend notifications into the
background of the stream.

They can be expanded, if one's curious, and clicking on an expanded toot will
jump there (assuming it is loaded).

This fixes #44 for the most part. Mention styling is left to be done.

Signed-off-by: Gergely Nagy <mastodon@gergo.csillger.hu>